+++ dunereco/RegCNN/func/RegCNNBoundary3D.h
@@ -22,14 +22,14 @@ namespace cnn
   public:
     /// Create new RegCNNBoundary3D object based on space points
     RegCNNBoundary3D(const int& nbinsX, const int& nbinsY, const int& nbinsZ, 
-             const float& x_length, const float& y_length, const float& z_length,
+             const double& x_length, const double& y_length, const double& z_length,
              const float& center_X, const float& center_Y, const float& center_Z);
 
     RegCNNBoundary3D(){};
 
     bool IsWithin(const float& rel_x, const float& rel_y, const float& rel_z);
 
-    float Length(const unsigned int& axis) const {return fStop[axis]-fStart[axis];};
+    double Length(const unsigned int& axis) const {return fStop[axis]-fStart[axis];};
     int NBins(const unsigned int& axis) const {return fStopBin[axis]-fStartBin[axis];};
     float StartPos(const unsigned int& axis) const {return fStart[axis];};
     float StopPos(const unsigned int& axis) const {return fStop[axis];};
@@ -37,8 +37,8 @@ namespace cnn
 
     int fStartBin[3]; ///< Minimum TDC in each view, inclusive
     int fStopBin[3];  ///< Maximum TDC in each view, inclusive
-    float fStart[3];  ///< Minimum wire, inclusive
-    float fStop[3];   ///< Maximum wire, inclusive
+    double fStart[3];  ///< Minimum wire, inclusive
+    double fStop[3];   ///< Maximum wire, inclusive
     float fCenter[3];
 
   };
